Dashboard populated :       show total profit lost , total sales, total cost and         tax properly iot company       show actuel currencies iot Open Curency Api       show latest three invoices iot company

+++ src/main/java/com/example/accountingapp/dto/InvoiceProductDTO.java
@@ -22,7 +22,7 @@ public class InvoiceProductDTO {
 
     @NotBlank
     @Size(min = 0)
-    private int qty;
+    private BigDecimal qty;
 
     @NotBlank
     @Size(min = 0)
