+++ src/main/java/com/example/accountingapp/entity/Product.java
@@ -8,6 +8,7 @@ import lombok.Setter;
 import org.hibernate.annotations.Where;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 @Entity
@@ -27,13 +28,13 @@ public class Product extends BaseEntity {
    private Category category;
 
 
-    private BigInteger qty;
+    private BigDecimal qty;
 
     @Enumerated(EnumType.STRING)
     private Unit unit;
 
-    private BigInteger lowLimitAlert;
-    private BigInteger tax;
+    private BigDecimal lowLimitAlert;
+    private BigDecimal tax;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "company_id")
